Work_Log_Generator/requirements.txt
# WorkLog
pygithub
PyQt5
qdarkstyle
